# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Mitovski Family Tree — Internal API
  version: 0.1.0
  description: |
    Internal HTTP API for the Mitovski family-tree project (idea.md §16). It is served
    on the Oracle VM behind Caddy and is called **only** by the Vercel Next.js BFF —
    never by browsers directly.

    ## Authentication (HMAC service auth, idea.md §4)

    Every `/v1/internal/*` operation requires an HMAC-signed request. The BFF sends
    these headers:

    - `X-Service-Id` — service identity
    - `X-Request-Timestamp` — RFC3339 UTC; rejected if skew > ±5 min
    - `X-Request-Nonce` — UUIDv4, single-use (replay protection)
    - `X-Idempotency-Key` — UUIDv4 on mutating requests
    - `X-Body-SHA256` — lowercase hex SHA-256 of the raw body
    - `X-Actor-Id` — acting principal (e.g. admin email, or `public`)
    - `X-Actor-Role` — `admin` | `public`; trusted **only** because it is signed
    - `X-Signature` — `hex(HMAC-SHA256(SERVICE_HMAC_SECRET, canonicalPayload))`

    Canonical payload (newline-joined, in order): HTTP method, request path, timestamp,
    nonce, idempotency key, body SHA-256, actor id, actor role. Exact escaping/empty-value
    rules and Go↔TypeScript test vectors live in `contracts/hmac.md` and
    `contracts/hmac-test-vectors.json` (Task 07).

    The reusable header parameters are defined in `components/parameters` and are attached
    to representative operations; **all** `/v1/internal/*` operations require the full set
...

...

...

    ## Authorization

...
servers:
...
security:
  - ServiceHmac: []
paths:
...
  /v1/internal/submission-people/{id}/create-person:
    post:
      operationId: postSubmissionPeopleByIdCreatePerson
      tags: [matching]
      summary: Promote a submitted person to a new canonical person (admin)
      description: Creates person, names, honest-precision events, and questionnaire source/evidence in one transaction (idea.md §7, §21).
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "201":
          description: New canonical person aggregate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Person" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /v1/internal/submission-people/{id}/link-person:
    post:
      operationId: postSubmissionPeopleByIdLinkPerson
      tags: [matching]
      summary: Link a submitted person to an existing canonical person (admin)
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [personId]
              properties:
                personId: { type: string, format: uuid }
      responses:
        "200":
          description: Linked person aggregate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Person" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/internal/people:
    get:
      operationId: getPeople
      tags: [people]
      summary: Search canonical people (admin)
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: livingStatus
          in: query
          schema: { $ref: "#/components/schemas/LivingStatus" }
        - name: privacyLevel
          in: query
          schema: { $ref: "#/components/schemas/PrivacyLevel" }
        - name: includeMerged
          in: query
          schema: { type: boolean, default: false }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Matching people.
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/PersonSummary" }
                  total: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: postPeople
      tags: [people]
      summary: Create a canonical person manually (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePersonRequest" }
      responses:
        "201":
          description: Created person aggregate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Person" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/internal/people/{id}:
    get:
      operationId: getPeopleById
      tags: [people]
      summary: Get a canonical person aggregate (admin)
      description: A merged person returns a redirect envelope with `mergedIntoPersonId` and status 409.
      parameters:
        - $ref: "#/components/parameters/PathId"
        - name: view
          in: query
          schema: { $ref: "#/components/schemas/View" }
      responses:
        "200":
          description: Person aggregate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Person" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Person is merged; body carries the redirect target.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MergedRedirect" }
    patch:
      operationId: patchPeopleById
      tags: [people]
      summary: Update privacy, living status, or notes (admin)
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchPersonRequest" }
      responses:
        "200":
          description: Updated person aggregate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Person" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/internal/people/{id}/evidence:
    get:
      operationId: getPeopleByIdEvidence
      tags: [people]
      summary: List all evidence about a person (admin)
      description: Evidence across the person and its names/events, joined to source metadata (idea.md §14).
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Evidence items with source metadata.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        subjectType: { type: string }
                        subjectId: { type: string, format: uuid }
                        assertion: { type: string }
                        stance: { type: string, enum: [supports, disputes] }
                        confidence: { type: [integer, "null"] }
                        notes: { type: [string, "null"] }
                        sourceId: { type: string, format: uuid }
                        sourceTitle: { type: string }
                        sourceType: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/internal/people/{id}/merge:
    post:
      operationId: postPeopleByIdMerge
      tags: [people]
      summary: Merge a duplicate person into a target (admin, transactional)
      description: Fully transactional; aborts on self-edge, cycle, or conflict (idea.md §8, Task 25).
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MergeRequest" }
      responses:
        "200":
          description: Merged target aggregate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Person" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
components:
  securitySchemes:
    ServiceHmac:
      type: apiKey
      in: header
      name: X-Signature
      description: |
        HMAC-SHA256 service authentication. The `X-Signature` header carries
        `hex(HMAC-SHA256(SERVICE_HMAC_SECRET, canonicalPayload))`, but the scheme also
        requires the full header set (`X-Service-Id`, `X-Request-Timestamp`,
        `X-Request-Nonce`, `X-Idempotency-Key`, `X-Body-SHA256`, `X-Actor-Id`,
        `X-Actor-Role`) — see the API description and `contracts/hmac.md`.
  parameters:
...
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Authentication failed (generic message; idea.md §4).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Insufficient role.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflict (invalid transition, duplicate, or idempotency key reused with a different body).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UnprocessableEntity:
      description: Domain rule violation (e.g. cycle_detected, unknown_person).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable code (e.g. unauthorized, cycle_detected, rate_limited).
            message:
              type: string
              description: Safe, human-readable message; never leaks internals.
            requestId:
              type: string
              description: Correlation id echoed from X-Request-Id.
...
    LivingStatus:
      type: string
      enum: [living, deceased, unknown]
    PrivacyLevel:
      type: string
      enum: [private, family, public]
    NameType:
      type: string
      enum: [primary, birth, married, alias, nickname, transliterated]
    EventType:
      type: string
      enum: [birth, death, residence, migration, occupation, education]
    DatePrecision:
      type: string
      enum: [exact, month, year, approximate, range, unknown]
    ParentChildType:
      type: string
      enum: [biological, adoptive, step, foster, guardian, unknown]
    VerificationStatus:
      type: string
      enum: [proposed, confirmed, disputed, rejected]
    UnionType:
      type: string
      enum: [marriage, partnership, unknown]
...
    View:
      type: string
      enum: [admin, public]
      default: public

    # ----- Invites -----
...
    CreatePersonRequest:
      type: object
      required: [firstName, surname, livingStatus]
      properties:
        firstName: { type: string, maxLength: 100 }
        middleName: { type: [string, "null"], maxLength: 100 }
        surname: { type: string, maxLength: 100 }
        birthSurname: { type: [string, "null"], maxLength: 100 }
        nickname: { type: [string, "null"], maxLength: 100 }
        livingStatus: { $ref: "#/components/schemas/LivingStatus" }
        privacyLevel: { $ref: "#/components/schemas/PrivacyLevel" }
    PatchPersonRequest:
      type: object
      properties:
        privacyLevel: { $ref: "#/components/schemas/PrivacyLevel" }
        livingStatus: { $ref: "#/components/schemas/LivingStatus" }
        notes: { type: [string, "null"], maxLength: 4000 }
    MergeRequest:
      type: object
      required: [targetPersonId, reason]
      properties:
        targetPersonId: { type: string, format: uuid }
        reason: { type: string, maxLength: 2000 }
    MergedRedirect:
      type: object
      required: [mergedIntoPersonId]
      properties:
        mergedIntoPersonId: { type: string, format: uuid }
    PersonName:
      type: object
      required: [id, nameType, isPreferred]
      properties:
        id: { type: string, format: uuid }
        firstName: { type: [string, "null"] }
        middleName: { type: [string, "null"] }
        surname: { type: [string, "null"] }
        birthSurname: { type: [string, "null"] }
        nickname: { type: [string, "null"] }
        normalizedName: { type: [string, "null"] }
        transliteratedName: { type: [string, "null"] }
        nameType: { $ref: "#/components/schemas/NameType" }
        isPreferred: { type: boolean }
        sourceId: { type: [string, "null"], format: uuid }
    PersonEvent:
      type: object
      required: [id, eventType, datePrecision]
      properties:
        id: { type: string, format: uuid }
        eventType: { $ref: "#/components/schemas/EventType" }
        value: { type: [string, "null"] }
        placeId: { type: [string, "null"], format: uuid }
        placeLabel: { type: [string, "null"] }
        dateFrom: { type: [string, "null"], format: date }
        dateTo: { type: [string, "null"], format: date }
        yearFrom: { type: [integer, "null"] }
        yearTo: { type: [integer, "null"] }
        datePrecision: { $ref: "#/components/schemas/DatePrecision" }
    PersonSummary:
      type: object
      required: [id, label, living]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        birthYear: { type: [integer, "null"] }
        deathYear: { type: [integer, "null"] }
        living: { type: boolean }
        privacyLevel: { $ref: "#/components/schemas/PrivacyLevel" }
        merged: { type: boolean }
    Person:
      type: object
      required: [id, livingStatus, privacyLevel, names]
      properties:
        id: { type: string, format: uuid }
        livingStatus: { $ref: "#/components/schemas/LivingStatus" }
        privacyLevel: { $ref: "#/components/schemas/PrivacyLevel" }
        notes: { type: [string, "null"] }
        mergedIntoPersonId: { type: [string, "null"], format: uuid }
        deletedAt: { type: [string, "null"], format: date-time }
        names:
          type: array
          items: { $ref: "#/components/schemas/PersonName" }
        events:
          type: array
          items: { $ref: "#/components/schemas/PersonEvent" }
        parents:
          type: array
          items: { $ref: "#/components/schemas/ParentChildRelationship" }
        children:
          type: array
          items: { $ref: "#/components/schemas/ParentChildRelationship" }
        unions:
          type: array
          items: { $ref: "#/components/schemas/FamilyUnion" }
        sourceCount: { type: integer }

    # ----- Relationships -----
...
    ParentChildRelationship:
      type: object
      required: [id, parentId, childId, relationshipType, verificationStatus]
      properties:
        id: { type: string, format: uuid }
        parentId: { type: string, format: uuid }
        childId: { type: string, format: uuid }
        relationshipType: { $ref: "#/components/schemas/ParentChildType" }
        familyUnionId: { type: [string, "null"], format: uuid }
        verificationStatus: { $ref: "#/components/schemas/VerificationStatus" }
        confidence: { type: [integer, "null"] }
    FamilyUnion:
      type: object
      required: [id, unionType, partnerIds]
      properties:
        id: { type: string, format: uuid }
        unionType: { $ref: "#/components/schemas/UnionType" }
        partnerIds:
          type: array
          items: { type: string, format: uuid }
        childIds:
          type: array
          items: { type: string, format: uuid }

    # ----- Tree projection (idea.md §13) -----
...